% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbondate.R
\name{carbondate}
\alias{carbondate}
\title{Carbon date a URL/URI}
\usage{
carbondate(uri)
}
\arguments{
\item{uri}{URL/URI to carbon date}
}
\value{
data frame (tibble)
}
\description{
Given a URL/URI, use all available methods to gather date info and present
results in a data frame
}
\examples{
carbondate(
  "http://www.cnn.com/2011/10/28/living/ways-to-look-better-is/index.html"
)
}
