% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbondater-package.R
\docType{package}
\name{carbondater}
\alias{carbondater}
\alias{carbondater-package}
\title{Estimate the Age of Web Resources}
\description{
Methods are provided to read 'URL' metadata and scan web archives/use
web archive 'APIs' to determine the approxate age of the resource.
}
\author{
Bob Rudis (bob@rud.is)
}
