% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-date-from-url.R
\name{get_date_from_url}
\alias{get_date_from_url}
\title{Locate and retrieve publication date from a URL}
\usage{
get_date_from_url(uri)
}
\arguments{
\item{uri}{URL/URI to examine}
}
\description{
Many modern URLs have publication dates in the URL. This function
searches for common ones and returns it if found.
}
\examples{
get_date_from_url(
  "https://rud.is/b/2018/05/03/seventeen-minutes-from-tweet-to-package/"
)
}
