% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-earliest-memento.R
\name{get_earliest_mementos}
\alias{get_earliest_mementos}
\title{Retrieve the earliest mementos for a URL/URI}
\usage{
get_earliest_mementos(uri)
}
\arguments{
\item{uri}{URL/URI to search for}
}
\value{
data frame (tibble) or \code{NULL} if no mementos
}
\description{
Retrieve the earliest mementos for a URL/URI
}
\examples{
get_earliest_mementos("http://rapid7.com/")
}
