% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-earliest-pubdate.R
\name{get_earliest_pubdate}
\alias{get_earliest_pubdate}
\title{Retrieve the earliest "pubdate" for a URL}
\usage{
get_earliest_pubdate(uri)
}
\arguments{
\item{uri}{URL/URI to read}
}
\value{
data frame (tibble)
}
\description{
Scan URL content for metatdata indicating publication date. Failing that,
try to extract it from the URL itself.
}
\examples{
get_earliest_pubdate(
 "https://rud.is/b/2018/05/03/seventeen-minutes-from-tweet-to-package/"
)
}
