% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-last-modified.R
\name{get_last_modified}
\alias{get_last_modified}
\title{Retrieve the last-modified header}
\usage{
get_last_modified(url)
}
\arguments{
\item{uri}{URL/URI to read}
}
\value{
data frame (tibble) or \code{NULL} if no haeder
}
\description{
Retrieve the last-modified header
}
\examples{
get_last_modified("http://www.example.org/")
}
